+++ pkg/core/xds/resource.go
@@ -1,6 +1,7 @@
 package xds
 
 import (
+	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes"
 
 	envoy "github.com/envoyproxy/go-control-plane/envoy/api/v2"
@@ -34,4 +35,51 @@ func (rs ResourceList) ToDeltaDiscoveryResponse() *envoy.DeltaDiscoveryResponse 
 		})
 	}
 	return resp
+}
+
+// ResourceSet represents a set of generic xDS resources.
+type ResourceSet struct {
+	// we want to keep resources in the order they were added
+	resources []*Resource
+	// we want to prevent duplicates
+	typeToNamesIndex map[string]map[string]bool
+}
+
+func (s *ResourceSet) Contains(name string, resource ResourcePayload) bool {
+	names, ok := s.typeToNamesIndex[s.typeName(resource)]
+	if !ok {
+		return false
+	}
+	_, ok = names[name]
+	return ok
+}
+
+func (s *ResourceSet) Add(resources ...*Resource) *ResourceSet {
+	for _, resource := range resources {
+		if s.Contains(resource.Name, resource.Resource) {
+			continue
+		}
+		s.resources = append(s.resources, resource)
+		s.index(resource)
+	}
+	return s
+}
+
+func (s *ResourceSet) typeName(resource ResourcePayload) string {
+	return proto.MessageName(resource)
+}
+
+func (s *ResourceSet) index(resource *Resource) {
+	if s.typeToNamesIndex == nil {
+		s.typeToNamesIndex = map[string]map[string]bool{}
+	}
+	typeName := s.typeName(resource.Resource)
+	if s.typeToNamesIndex[typeName] == nil {
+		s.typeToNamesIndex[typeName] = map[string]bool{}
+	}
+	s.typeToNamesIndex[typeName][resource.Name] = true
+}
+
+func (s *ResourceSet) List() []*Resource {
+	return s.resources
 }
