+++ pkg/core/xds/resource_test.go
@@ -1,12 +1,10 @@
-package generator_test
+package xds_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	. "github.com/Kong/kuma/pkg/xds/generator"
-
-	core_xds "github.com/Kong/kuma/pkg/core/xds"
+	. "github.com/Kong/kuma/pkg/core/xds"
 
 	envoy "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 )
@@ -22,7 +20,7 @@ var _ = Describe("ResourceSet", func() {
 
 	It("set of 1 element should return a list of 1 element", func() {
 		// given
-		resource := &core_xds.Resource{
+		resource := &Resource{
 			Name:    "backend",
 			Version: "v1",
 			Resource: &envoy.Cluster{
@@ -40,14 +38,14 @@ var _ = Describe("ResourceSet", func() {
 
 	It("set of 2 elements should return a list of 2 elements", func() {
 		// given
-		resource1 := &core_xds.Resource{
+		resource1 := &Resource{
 			Name:    "backend",
 			Version: "v1",
 			Resource: &envoy.Cluster{
 				Name: "backend",
 			},
 		}
-		resource2 := &core_xds.Resource{
+		resource2 := &Resource{
 			Name:    "outbound:127.0.0.1:8080",
 			Version: "v2",
 			Resource: &envoy.Listener{
@@ -68,14 +66,14 @@ var _ = Describe("ResourceSet", func() {
 
 	It("should not be possible to add 2 resources with same name and type", func() {
 		// given
-		resource1 := &core_xds.Resource{
+		resource1 := &Resource{
 			Name:    "backend",
 			Version: "v1",
 			Resource: &envoy.Cluster{
 				Name: "backend",
 			},
 		}
-		resource2 := &core_xds.Resource{
+		resource2 := &Resource{
 			Name:    "backend",
 			Version: "v2",
 			Resource: &envoy.Cluster{
@@ -96,14 +94,14 @@ var _ = Describe("ResourceSet", func() {
 
 	It("should be possible to add 2 resources with same name but different types", func() {
 		// given
-		resource1 := &core_xds.Resource{
+		resource1 := &Resource{
 			Name:    "backend",
 			Version: "v1",
 			Resource: &envoy.Cluster{
 				Name: "backend",
 			},
 		}
-		resource2 := &core_xds.Resource{
+		resource2 := &Resource{
 			Name:    "backend",
 			Version: "v2",
 			Resource: &envoy.Listener{
